console: test that flow detail views don't crash
